Remove Anki compression step
Well, that didn't last long. GitHub has issues with files over 100MB,
and even the compressed version of my collection is over 200MB. So
instead of sending my collection to GitHub, I've uploaded the CSS file
that powers it, along with a basic explanation of how it works. Now
PyFile has a function for copying that CSS file over to the repository
folder so I don't have to do updates manually.