Starting work on the database/universal objects
Although this is only some rudimentary progress, this is my first step
towards utilizing both an object factory (started with the "factory"
class method) and a database file (created with sqlite3). It's not ready
yet, but I will replace the JSON file that is currently used with this
database file (settings.db) very soon. For now, I can only add database
records or update them and the factory is still under construction.