Continued work on database interaction
Getting closer to switching from JSON to a Sqlite3 database. In this
commit, I've added a backup_loop class method that will replace the for
loops that are customized in each DirObject instance. I've also included
a method for create a variable out of a database entry. Aside from these
changes, I also moved my local repository files around, and had to
change the path variable for the JSON file (which will hopefully be
obsolete soon!).